Fix needing to press login button twice by moving initGameSocket() from App.vue to HomePage.vue
