+++ containers/vuejs/src/getSocket.ts
@@ -1,14 +1,9 @@
 import { io, Socket } from 'socket.io-client'
 
 export let gameSocket: Socket
-export let chatSocket: Socket
 
 export const initGameSocket = (opts: any) => {
   gameSocket = getSocket('/game', opts)
-}
-
-export const initChatSocket = (opts: any) => {
-  chatSocket = getSocket('/chat', opts)
 }
 
 const getSocket = (namespace: string, opts: any) => {
